moved more of UI into new format
