+++ src/ATGUI/Windows/main.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "../atgui.h"
+
+namespace Main
+{
+	extern bool showWindow;
+
+	extern void RenderWindow();
+}
