+++ src/Hacks/radar.h
@@ -15,7 +15,7 @@ namespace Radar
 		SHAPE_TRIANGLE_UPSIDEDOWN
 	};
 	ImColor GetRadarPlayerColor(C_BasePlayer* player, bool visible);
-	void DrawWindow();
+	void RenderWindow();
 	void InGameRadar(C_BasePlayer* player);
 	void BeginFrame();
 };
